+++ CHANGELOG.md
@@ -5,11 +5,23 @@
 - [feature] Added the `firebase_admin.messaging` module for sending
   Firebase notifications and managing topic subscriptions. 
 
+### Authentication
+
+- [added] The ['verify_id_token()'](https://firebase.google.com/docs/reference/admin/python/firebase_admin.auth#verify_id_token)
+  function now accepts an optional `check_revoked` parameter. When `True`, an 
+  additional check is performed to see whether the token has been revoked. 
+- [added] A new
+  ['auth.revoke_refresh_tokens(uid)'](https://firebase.google.com/docs/reference/admin/python/firebase_admin.auth#revoke_refresh_tokens)
+  function has been added to invalidate all tokens issued to a user.
+- [added] A new `tokens_valid_after_timestamp` property has been added to the   
+  ['UserRecord'](https://firebase.google.com/docs/reference/admin/python/firebase_admin.auth#userrecord),
+  class indicating the time before which tokens are not valid.
+
 # v2.8.0
 
 ### Initialization
 
-- [added] The [`initialize_app()`](https://firebase.google.com/docs/reference/admin/node/admin#.initializeApp)
+- [added] The [`initialize_app()`](https://firebase.google.com/docs/reference/admin/python/firebase_admin#initialize_app)
   method can now be invoked without any arguments. This initializes an app
   using Google Application Default Credentials, and other
   options loaded from the `FIREBASE_CONFIG` environment variable.
@@ -30,6 +42,7 @@
 # v2.6.0
 
 ### Authentication
+
 - [added] Added the
   [`list_users()`](https://firebase.google.com/docs/reference/admin/python/firebase_admin.auth#list_users)
   function to the `firebase_admin.auth` module. This function enables listing
@@ -41,6 +54,7 @@
   user's ID token.
 
 ### Realtime Database
+
 - [changed] Updated the `start_at()`, `end_at()` and `equal_to()` methods of
   the [`db.Query`](https://firebase.google.com/docs/reference/admin/python/firebase_admin.db#query) class
   so they can accept empty string arguments.